Refactor mixing-plane interpolator for conciseness + remove unused vars & functions

+++ Common/include/interface_interpolation/CInterpolator.hpp
@@ -102,7 +102,7 @@ class CInterpolator {
   vector<vector<CDonorInfo>> targetVertices; /*! \brief Donor information per marker per vertex of the target. */
 
   struct CSpanDonorInfo {
-    unsigned long donorSpan;  // Refers to donor span
+    size_t donorSpan;  // Refers to donor span
     su2double coefficient;    // Refers to coefficient
   };
   vector<vector<CSpanDonorInfo>> targetSpans;  // <iMarkerInt<TargetSpan<CSpanDonorInfo>>
